Lock cosmic entity sculptures behind Forbidden_Sculptures tech.
The cosmic entity sculptures previously had no `researchPrerequisites`
listed, meaning they could be built from the start, removing the need
for players to research `Forbidden_Sculptures` at all.

This commit creates an abstract `MonstrousSculpture` base with the
research lock, along with other common aspects between the sculptures.

This also standardizes `WorkToMake` to 90000 for each major sculpture,
rather than a couple of sculptures being slightly more expensive (95000
or 96000).

The `YigSculptureLarge` is not changed by this commit, although I did
adjust the spacing of an XML tag to make it more pleasing to entities
from beyond the stars.